Added a Test class for ChairWarehouse
