+++ src/main/java/com/asgarov/Main.java
@@ -1,0 +1,13 @@
+package com.asgarov;
+
+public class Main {
+
+    public static void main(String[] args) {
+//        for (int i = 0; i < 10; i++) {
+//            new Thread(() -> {
+//                ChairWarehouse.getInstance();
+//                ChairWarehouse.getInstance().sellChairFromStock(1);
+//            }).start();
+//        }
+    }
+}
